Add and use a new wfe judgment in the Redex model.
In Martin's words:

I thought of another possibility to fix DOT: Add a new judgement

 Gamma |- T wfe

(for well-formed and expanding). A single rule for it:

 Gamma |- T wf    Gamma |- T \expand_z Ds
 ----------------------------------------------------------
 Gamma |- T wfe

Change the preconditions of all wf judgements so that all occurrences
of wf are replaced by wfe. That should guarantee that a well formed
type expands everywhere. Then, I believe we get transititivity of
subtyping for wfe types. Also change the two occurrences of wf in
typing judgements to wfe. This hopefully means that we need to use
subtyping only on wfe types.