Getting started with end-to-end tests.
Had to rename Success and Failure in TyperMonad, because it conflicts
with the same classes in Scala's parser combinators.

Also for test1() in TestDot to pass, had to change wf so that the two
alternatives for type selections are not both taken, since we want to
maintain uniqueness of the result.