Fix LR proof counterex in sel case.
Suppose context has

a: Top {z => i: Top {z => X: Bot .. Top { l: x.Z }; x.Z }

Now, we used to be able to complete the store and context as follows:

b = Top { z => X: Top .. Top; l: z.X } (l=b)
a = ... (i=b).

This is not good, because b doesn't satisfy what a expects of its i
field. The problem is that we didn't check that the b.X bounds are at
least as tight as what is specified.

Now, we do.